+++ src/Core/CostControl.Application/Features/MonetaryFund/Handlers/Commands/DeleteMonetaryFundCommandHandler.cs
@@ -19,6 +19,14 @@ namespace CostControl.Application.Features.MonetaryFund.Handlers.Commands
             var response = new BaseCommandResponse();
             try
             {
+                if (await _unitOfWork.DepositRepository.DepositByMonetaryFundExists(request.Id) || 
+                    await _unitOfWork.ExpenseRepository.ExpenseByMonetaryFundExist(request.Id))
+                {
+                    response.Success = false;
+                    response.Message = "No se puede eliminar este fondo monetario porque está vinculado a uno o más Depósitos o Gastos.";
+                    return response;
+                }
+
                 await _unitOfWork.MonetaryFundRepository.DeleteAsync(request.Id);
                 await _unitOfWork.SaveAsync();
             }
